+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- UI: Dev info panel to room setttings and message information panel
+
 ### Changed
 
 - UI: Adjusted which rooms are shows as public
+- UI: Show application version in About Panel
+- UI: Fix typo in "Change Power Levels" for event type
 - SDK: Use unread marker calculation from Trixnity
 
 ### Deprecated
@@ -45,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support UnifiedPush
+- SDK: Support UnifiedPush
 
 ### Changed